requirements.txt
PyYAML>=6.0
requests>=2.28.0
docker>=6.0.0
paramiko>=3.0.0
Flask>=2.0.0